+++ packages/cinterop/src/darwin/kotlin/io/realm/internal/interop/RealmInterop.kt
@@ -173,6 +173,13 @@ fun realm_value_t.set(memScope: MemScope, value: Any?): realm_value_t {
             type = realm_value_type.RLM_TYPE_DOUBLE
             dnum = value
         }
+        is Timestamp -> {
+            type = realm_value_type.RLM_TYPE_TIMESTAMP
+            timestamp.apply {
+                seconds = value.seconds
+                nanoseconds = value.nanoSeconds
+            }
+        }
         else ->
             TODO("Value conversion not yet implemented for : ${value::class.simpleName}")
     }
@@ -393,7 +400,7 @@ actual object RealmInterop {
             val outCount = alloc<size_tVar>()
             checkedBooleanResult(realm_wrapper.realm_get_class_keys(realm.cptr(), keys, max.convert(), outCount.ptr))
             if (max != outCount.value.toLong()) {
-                error("Invalid schema: Insufficient keys; got ${outCount.value} expected $max")
+                error("Invalid schema: Insufficient keys; got ${outCount.value}, expected $max")
             }
             return (0 until max).map { ClassKey(keys[it].toLong()) }
         }
@@ -458,8 +465,8 @@ actual object RealmInterop {
                             PropertyInfo(
                                 name.safeKString("name"),
                                 public_name?.toKString(),
-                                PropertyType.of(type.toInt()),
-                                CollectionType.of(collection_type.toInt()),
+                                PropertyType.from(type.toInt()),
+                                CollectionType.from(collection_type.toInt()),
                                 link_target?.toKString(),
                                 link_origin_property_name?.toKString(),
                                 PropertyKey(key),
@@ -579,6 +586,8 @@ actual object RealmInterop {
                 value.fnum
             realm_value_type.RLM_TYPE_DOUBLE ->
                 value.dnum
+            realm_value_type.RLM_TYPE_TIMESTAMP ->
+                value.asTimestamp()
             realm_value_type.RLM_TYPE_LINK ->
                 value.asLink()
             else ->
@@ -671,7 +680,7 @@ actual object RealmInterop {
         return realm_wrapper.realm_list_is_valid(list.cptr())
     }
 
-    @Suppress("ComplexMethod")
+    @Suppress("ComplexMethod", "LongMethod")
     private fun <T> MemScope.to_realm_value(value: T): realm_value_t {
         val cvalue: realm_value_t = alloc()
         when (value) {
@@ -713,6 +722,13 @@ actual object RealmInterop {
             is Double -> {
                 cvalue.type = realm_value_type.RLM_TYPE_DOUBLE
                 cvalue.dnum = value as Double
+            }
+            is Timestamp -> {
+                cvalue.type = realm_value_type.RLM_TYPE_TIMESTAMP
+                cvalue.timestamp.apply {
+                    seconds = value.seconds
+                    nanoseconds = value.nanoSeconds
+                }
             }
             is RealmObjectInterop -> {
                 cvalue.type = realm_value_type.RLM_TYPE_LINK
@@ -1259,6 +1275,13 @@ actual object RealmInterop {
             )
         )
         return propertyInfo
+    }
+
+    private fun realm_value_t.asTimestamp(): Timestamp {
+        if (this.type != realm_value_type.RLM_TYPE_TIMESTAMP) {
+            error("Value is not of type Timestamp: $this.type")
+        }
+        return TimestampImpl(this.timestamp.seconds, this.timestamp.nanoseconds)
     }
 
     private fun realm_value_t.asLink(): Link {