+++ packages/library-base/src/commonMain/kotlin/io/realm/internal/schema/RealmClassImpl.kt
@@ -20,22 +20,22 @@ import io.realm.internal.interop.ClassInfo
 import io.realm.internal.interop.PropertyInfo
 import io.realm.schema.RealmClass
 import io.realm.schema.RealmProperty
-import io.realm.schema.SingularPropertyType
+import io.realm.schema.ValuePropertyType
 
 data class RealmClassImpl(
     // Optimization: Store the schema in the C-API alike structure directly from compiler plugin to
     // avoid unnecessary repeated initializations for realm_schema_new
-    val cinterosClass: ClassInfo,
+    val cinteropClass: ClassInfo,
     val cinteropProperties: List<PropertyInfo>
 ) : RealmClass {
 
-    override val name: String = cinterosClass.name
+    override val name: String = cinteropClass.name
     override val properties: Collection<RealmProperty> = cinteropProperties.map {
         RealmPropertyImpl.fromCoreProperty(it)
     }
 
     override fun get(key: String): RealmProperty? = properties.firstOrNull { it.name == key }
     override fun primaryKey(): RealmProperty? = properties.firstOrNull {
-        it.type.run { this is SingularPropertyType && isPrimaryKey }
+        it.type.run { this is ValuePropertyType && isPrimaryKey }
     }
 }