+++ packages/plugin-compiler/src/test/kotlin/io/realm/compiler/GenerationExtensionTest.kt
@@ -72,12 +72,11 @@ class GenerationExtensionTest {
         ).joinToString(separator = File.separator)
 
         fun assertGeneratedIR() {
-            // TODO Quick fix: Investigate where 'main' part suddenly came from with Kotlin 1.5.31
-            stripInputPath(File("${outputDir()}/main/00_ValidateIrBeforeLowering.ir"), fileMap)
+            val outputFile = File("${outputDir()}/main/00_ValidateIrBeforeLowering.ir")
+            stripInputPath(outputFile, fileMap)
             assertEquals(
                 File("${expectedDir()}/00_ValidateIrBeforeLowering.ir").readText(),
-                // TODO Quick fix: Investigate where 'main' part suddenly came from with Kotlin 1.5.31
-                File("${outputDir()}/main/00_ValidateIrBeforeLowering.ir").readText()
+                outputFile.readText()
             )
         }
     }
@@ -200,6 +199,7 @@ class GenerationExtensionTest {
             "booleanField" to PropertyType.RLM_PROPERTY_TYPE_BOOL,
             "floatField" to PropertyType.RLM_PROPERTY_TYPE_FLOAT,
             "doubleField" to PropertyType.RLM_PROPERTY_TYPE_DOUBLE,
+            "timestampField" to PropertyType.RLM_PROPERTY_TYPE_TIMESTAMP,
 
             // RealmObject
             "child" to PropertyType.RLM_PROPERTY_TYPE_OBJECT,
@@ -214,6 +214,7 @@ class GenerationExtensionTest {
             "booleanListField" to PropertyType.RLM_PROPERTY_TYPE_BOOL,
             "floatListField" to PropertyType.RLM_PROPERTY_TYPE_FLOAT,
             "doubleListField" to PropertyType.RLM_PROPERTY_TYPE_DOUBLE,
+            "timestampListField" to PropertyType.RLM_PROPERTY_TYPE_TIMESTAMP,
             "objectListField" to PropertyType.RLM_PROPERTY_TYPE_OBJECT,
 
             // Nullable list types
@@ -225,7 +226,8 @@ class GenerationExtensionTest {
             "nullableLongListField" to PropertyType.RLM_PROPERTY_TYPE_INT,
             "nullableBooleanListField" to PropertyType.RLM_PROPERTY_TYPE_BOOL,
             "nullableFloatListField" to PropertyType.RLM_PROPERTY_TYPE_FLOAT,
-            "nullableDoubleListField" to PropertyType.RLM_PROPERTY_TYPE_DOUBLE
+            "nullableDoubleListField" to PropertyType.RLM_PROPERTY_TYPE_DOUBLE,
+            "nullableTimestampListField" to PropertyType.RLM_PROPERTY_TYPE_TIMESTAMP,
         )
         assertEquals(expectedProperties.size, properties.size)
         properties.map { property ->