+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 * None.
 
 ### Enhancements
-* Added support for `User.logOut()` ([#245](https://github.com/realm/realm-kotlin/issues/245)).
+* [Sync] Added support for `User.logOut()` ([#245](https://github.com/realm/realm-kotlin/issues/245)).
+* Added supported for dates through a new property type: `RealmInstant`.
 
 ### Fixed
 * Gradle metadata for pure Android projects. Now using `io.realm.kotlin:library-base:<VERSION>` should work correctly.